+++ whats-open/api/models.py
@@ -8,7 +8,7 @@ the API.
 
 https://docs.djangoproject.com/en/1.11/topics/db/models/
 """
-# Python stdlib Imports
+# Python Imports
 import datetime
 
 # Django Imports
@@ -22,7 +22,7 @@ from django.utils import timezone
 from model_utils.models import TimeStampedModel
 from autoslug import AutoSlugField
 from taggit.managers import TaggableManager
-from taggit.models import GenericTaggedItemBase, TagBase
+
 
 class Category(TimeStampedModel):
     """
@@ -34,6 +34,7 @@ class Category(TimeStampedModel):
     - Gyms
     - Study areas (Libraries, The Ridge, JC, etc)
     """
+
     # The name of the category
     name = models.CharField(max_length=100)
 
@@ -41,7 +42,7 @@ class Category(TimeStampedModel):
         verbose_name = "category"
         verbose_name_plural = "categories"
         # Sort by name in admin view.
-        ordering = ['name']
+        ordering = ["name"]
 
     def __str__(self):
         """
@@ -49,27 +50,29 @@ class Category(TimeStampedModel):
         """
         return self.name
 
+
 class Location(TimeStampedModel):
     """
     Represents a specific location that a Facility can be found.
     """
+
     CAMPUS_LOCATIONS = (
-        # (set in model, human readable version)
+        ("front royal", "Front Royal"),
         ("prince william", "Prince William County Science and Technology"),
-        ("korea", "Mason Korea"),
         ("fairfax", "Fairfax"),
-        ("arlington", "Arlington")
+        ("arlington", "Arlington"),
     )
     # The building that the facility is located in (on campus).
     building = models.CharField(max_length=100)
-    friendly_building = models.CharField('Building Abbreviation',
-                                         help_text="Example: Exploratory Hall becomes EXPL",
-                                         blank=True,
-                                         max_length=10)
+    friendly_building = models.CharField(
+        "Building Abbreviation",
+        help_text="Example: Exploratory Hall becomes EXPL",
+        blank=True,
+        max_length=10,
+    )
     # The physical address of the facility.
     address = models.CharField(max_length=100)
-    campus_region = models.CharField(choices=CAMPUS_LOCATIONS,
-                                     max_length=100, default="fairfax")
+    campus_region = models.CharField(choices=CAMPUS_LOCATIONS, max_length=100)
     # Boolean for whether or not the location is "on campus" or not.
     on_campus = models.BooleanField(default=True)
     # A GeoJson coordinate pair that represents the physical location
@@ -83,18 +86,8 @@ class Location(TimeStampedModel):
         """
         String representation of a Location object.
         """
-        return 'Found in %s at %s | On Campus: %s' % (self.building,
-                                                      self.address,
-                                                      self.on_campus)
+        return "%s on %s Campus" % (self.building, self.campus_region.title())
 
-# Look I didn't want to do this but APPARENTLY you cannot have two
-# TaggableManager()s on a model and thus you have to make a WHOLE other model
-# to have this work.
-# https://neutron-drive.appspot.com/blog/multiple-tags
-class StupidFacilityLabelHack(TagBase):
-    pass
-class StupidLabelHack(GenericTaggedItemBase):
-    tag = models.ForeignKey(StupidFacilityLabelHack, on_delete=models.CASCADE, related_name="%(app_label)s_%(class)s_items")
 
 class Facility(TimeStampedModel):
     """
@@ -102,73 +95,96 @@ class Facility(TimeStampedModel):
     establishment that has a schedule of open hours and a location that serves
     a specific purpose that can be categorized.
     """
+
     # The name of the Facility
     facility_name = models.CharField(max_length=100)
     # Instead of id
-    slug = AutoSlugField(populate_from='facility_name', unique=True)
+    slug = AutoSlugField(populate_from="facility_name", unique=True)
 
     # The category that this facility can be grouped with
-    facility_category = models.ForeignKey('Category',
-                                          related_name="categories",
-                                          on_delete=models.CASCADE)
+    facility_category = models.ForeignKey(
+        "Category", related_name="categories", on_delete=models.CASCADE
+    )
     # The location object that relates to this facility
-    facility_location = models.ForeignKey('Location',
-                                          related_name="facilities",
-                                          on_delete=models.CASCADE)
+    facility_location = models.ForeignKey(
+        "Location", related_name="facilities", on_delete=models.CASCADE
+    )
 
     # A note that can be left on a Facility to provide some additional
     # information.
-    note = models.TextField('Facility Note', blank=True,
-                            help_text="Additional information that is sent with this Facility.")
+    note = models.TextField(
+        "Facility Note",
+        blank=True,
+        help_text="Additional information that is sent with this Facility.",
+    )
 
     # A link to the logo image for this Facility
-    logo = models.URLField('Logo URL', blank=True,
-                           default="https://imgur.com/q2Phkn9.png",
-                           help_text="The absolute URL to the logo image for this Facility.")
+    logo = models.URLField(
+        "Logo URL",
+        blank=True,
+        default="https://wopen-cdn.dhaynes.xyz/default.png",
+        help_text="The absolute URL to the logo image for this Facility.",
+    )
 
     # The User(s) that claim ownership over this facility
     owners = models.ManyToManyField(User)
 
     # The schedule that is defaulted to if no special schedule is in effect
-    main_schedule = models.ForeignKey('Schedule',
-                                      related_name='facility_main',
-                                      on_delete=models.CASCADE)
+    main_schedule = models.ForeignKey(
+        "Schedule", related_name="facility_main", on_delete=models.CASCADE
+    )
     # A schedule that has a specific start and end date
-    special_schedules = models.ManyToManyField('Schedule',
-                                               related_name='facility_special',
-                                               blank=True,
-                                               help_text="This schedule will come into effect only for its specified duration.")
+    special_schedules = models.ManyToManyField(
+        "Schedule",
+        related_name="facility_special",
+        blank=True,
+        help_text="This schedule will come into effect only for its specified duration.",
+    )
 
     # URL, if it exists, to the Tapingo page that is associated with this
     # facility
-    tapingo_url = models.URLField(blank=True, validators=[RegexValidator(regex='^https:\/\/www.tapingo.com\/',
-                                                                         message='The link is not a valid tapingo link. Example: https://www.tapingo.com/order/restaurant/starbucks-gmu-johnson/',
-                                                                         code='invalid_tapingo_url')])
+    tapingo_url = models.URLField(
+        blank=True,
+        validators=[
+            RegexValidator(
+                regex="^https:\/\/www.tapingo.com\/",
+                message="The link is not a valid tapingo link. Example: https://www.tapingo.com/order/restaurant/starbucks-gmu-johnson/",
+                code="invalid_tapingo_url",
+            )
+        ],
+    )
 
     # Phone number for a location if provided. Accept both ###-###-#### or
     # without dashes
-    phone_number = models.CharField(blank=True, max_length=18, validators=[RegexValidator(regex='^\(?([0-9]{3})\)?[-.●]?([0-9]{3})[-.●]?([0-9]{4})$',
-                                                                           message='Invalid phone number',
-                                                                           code='invalid_phone_number')])
+    phone_number = models.CharField(
+        blank=True,
+        max_length=18,
+        validators=[
+            RegexValidator(
+                regex="^\(?([0-9]{3})\)?[-.●]?([0-9]{3})[-.●]?([0-9]{4})$",
+                message="Invalid phone number",
+                code="invalid_phone_number",
+            )
+        ],
+    )
 
     # A comma seperate list of words that neatly and aptly describe the product
     # that this facility produces. (ex. for Taco Bell: mexican, taco, cheap)
     # These words are not shown to the use but are rather used in search.
-    facility_product_tags = TaggableManager(related_name="product_tags", help_text="A comma seperate list of words that neatly and aptly describe the product that this facility produces. These words are not shown to the use but are rather used in search.")
-
-    # Labels to describe the Facility that are displayed to the user and can be
-    # informative. "Takes Mason Money"
-    facility_labels = TaggableManager("labels", related_name="labels", through=StupidLabelHack, help_text="Labels to describe the Facility that are displayed to the user and can be informative.")
-
-    # Tag a Facility to be shown on the ShopMason or Sodoxo (or both)
-    # What's Open sites.
-    FACILITY_CLASSES = (
-        ("shopmason", "shopMason Facility"),
-        ("sodoxo", "Sodoxo Facility")
+    facility_product_tags = TaggableManager(
+        related_name="product_tags",
+        help_text="A comma seperate list of words that neatly and aptly describe the product that this facility produces. These words are not shown to the use but are rather used in search.",
     )
-    facility_classifier = models.CharField(choices=FACILITY_CLASSES,
-                                           help_text="Tag this facility to be shown on the ShopMason or Sodoxo What's Open sites.",
-                                           max_length=100, blank=True)
+
+    # Tag a Facility to be shown on the ShopMason
+    # What's Open sites.
+    FACILITY_CLASSES = (("shopmason", "shopMason Facility"),)
+    facility_classifier = models.CharField(
+        choices=FACILITY_CLASSES,
+        help_text="Tag this facility to be shown on the ShopMason What's Open sites.",
+        max_length=100,
+        blank=True,
+    )
 
     def is_open(self):
         """
@@ -177,49 +193,22 @@ class Facility(TimeStampedModel):
         First checks any valid special schedules and then checks the main,
         default, schedule.
         """
-        # Get the current date
-        today = datetime.datetime.today().date()
-        # Check special schedules first, loop through all of them
+        today = timezone.now()
+        # Check special schedules first
         for schedule in self.special_schedules.all():
             # Special schedules must have valid_start and valid_end set
             if schedule.valid_start and schedule.valid_end:
-                # If a special schedule in in effect
                 if schedule.valid_start <= today <= schedule.valid_end:
-                    # Check if the facility is open or not based on that
-                    # special schedule
-                    if schedule.is_open_now():
-                        # Open
-                        return True
-                    else:
-                        # Closed
-                        return False
-        # If no special schedule is in effect then check if the facility is
-        # open using the main_schedule
-        if self.main_schedule.is_open_now():
-            # Open
-            return True
-        else:
-            # Closed
-            return False
+                    return schedule.is_open_now()
 
-    def clean_schedules(self):
-        """
-        Loop through every special_schedule and remove entries that have
-        expired as well as promote special schedules to main if necessary.
-        """
-        for special_schedule in self.special_schedules.all():
-            # If it ends before today
-            if special_schedule.valid_end < timezone.now()  and special_schedule.schedule_for_removal:
-                self.special_schedules.remove(special_schedule)
-            elif special_schedule.promote_to_main:
-                if special_schedule.valid_start < timezone.now() and special_schedule.valid_end >= timezone.now():
-                    self.main_schedule = special_schedule
+        # If no special schedule is in effect then check the main_schedule
+        return self.main_schedule.is_open_now()
 
     class Meta:
         verbose_name = "facility"
         verbose_name_plural = "facilities"
         # Sort by name in admin view
-        ordering = ['facility_name']
+        ordering = ["facility_name"]
 
     def __str__(self):
         """
@@ -227,36 +216,40 @@ class Facility(TimeStampedModel):
         """
         return self.facility_name
 
+
 class Schedule(TimeStampedModel):
     """
     A period of time between two dates that represents the beginning and end of
     a "schedule" or rather, a collection of open times for a facility.
     """
+
     # The name of the schedule
     name = models.CharField(max_length=100)
 
     # The start date of the schedule
     # (inclusive)
-    valid_start = models.DateTimeField('Start Date', null=True, blank=True,
-                                       help_text="Date & time that this schedule goes into effect")
+    valid_start = models.DateTimeField(
+        "Start Date",
+        null=True,
+        blank=True,
+        help_text="Date & time that this schedule goes into effect",
+    )
     # The end date of the schedule
     # (inclusive)
-    valid_end = models.DateTimeField('End Date', null=True, blank=True,
-                                     help_text="Last date & time that this schedule is in effect")
+    valid_end = models.DateTimeField(
+        "End Date",
+        null=True,
+        blank=True,
+        help_text="Last date & time that this schedule is in effect",
+    )
 
     # Boolean for if this schedule is 24 hours
-    twenty_four_hours = models.BooleanField('24 hour schedule?', blank=True,
-                                            default=False, help_text="Toggle to True if the Facility is open 24 hours. You do not need to specify any Open Times, it will always be displayed as open.")
-
-    # Boolean for if this schedule should never be removed.
-    schedule_for_removal = models.BooleanField('Schedule for removal?',
-                                               blank=False, default=True,
-                                               help_text="Toggle to False if the schedule should never be removed in the backend. By default, all schedules are automatically deleted after they have expired.")
-    # Boolean for if this schedule should become the main schedule at the point
-    # it goes live
-    promote_to_main = models.BooleanField('Schedule for promotion?',
-                                          blank=False, default=False,
-                                          help_text="Upon the start of the schedule, it will be promoted to become the main schedule of the Facility it is attached to rather than a special schedule.")
+    twenty_four_hours = models.BooleanField(
+        "24 hour schedule?",
+        blank=True,
+        default=False,
+        help_text="Toggle to True if the Facility is open 24 hours. You do not need to specify any Open Times, it will always be displayed as open.",
+    )
 
     def is_open_now(self):
         """
@@ -265,21 +258,17 @@ class Schedule(TimeStampedModel):
         # If the schedule is a 24 hour one, then it's open.
         if self.twenty_four_hours:
             return True
-        # Otherwise let's check if it's open.
         else:
             # Loop through all the open times that correspond to this schedule
             for open_time in OpenTime.objects.filter(schedule=self):
-                # If the current time we are looking at is open, then the schedule
-                # will say that the facility is open
                 if open_time.is_open_now():
-                    # Open
                     return True
             # Closed (all open times are not open)
             return False
 
     class Meta:
         # Sort by name in admin view
-        ordering = ['name']
+        ordering = ["name"]
 
     def __str__(self):
         """
@@ -294,6 +283,7 @@ class OpenTime(TimeStampedModel):
 
     Monday = 0, Sunday = 6.
     """
+
     # Define integer constants to represent days of the week
     MONDAY = 0
     TUESDAY = 1
@@ -305,18 +295,19 @@ class OpenTime(TimeStampedModel):
 
     # Tuple that ties a day of the week with an integer representation
     DAY_CHOICES = (
-        (MONDAY, 'Monday'),
-        (TUESDAY, 'Tuesday'),
-        (WEDNESDAY, 'Wednesday'),
-        (THURSDAY, 'Thursday'),
-        (FRIDAY, 'Friday'),
-        (SATURDAY, 'Saturday'),
-        (SUNDAY, 'Sunday'),
+        (MONDAY, "Monday"),
+        (TUESDAY, "Tuesday"),
+        (WEDNESDAY, "Wednesday"),
+        (THURSDAY, "Thursday"),
+        (FRIDAY, "Friday"),
+        (SATURDAY, "Saturday"),
+        (SUNDAY, "Sunday"),
     )
 
     # The schedule that this period of open time is a part of
-    schedule = models.ForeignKey('Schedule', related_name='open_times',
-                                 on_delete=models.CASCADE)
+    schedule = models.ForeignKey(
+        "Schedule", related_name="open_times", on_delete=models.CASCADE
+    )
 
     # The day that the open time begins on
     start_day = models.IntegerField(default=0, choices=DAY_CHOICES)
@@ -334,6 +325,7 @@ class OpenTime(TimeStampedModel):
         """
         # Get the current datetime
         today = datetime.datetime.today()
+
         # Check that the start occurs before the end
         if self.start_day <= self.end_day:
             # If today is the start_day
@@ -356,7 +348,7 @@ class OpenTime(TimeStampedModel):
             elif self.end_day < today.weekday():
                 # Closed
                 return False
-        # The end_day > start_day
+        # The start_day > end_day
         else:
             # If today is the start_day
             if self.start_day == today.weekday():
@@ -375,20 +367,30 @@ class OpenTime(TimeStampedModel):
             if self.end_day < today.weekday() < self.start_day:
                 # Closed
                 return False
-        # All checks passed, it's Open
+        # All checks passed, it's open
         return True
 
     def __str__(self):
         """
         String representation of a OpenTime object.
         """
-        weekdays = ['Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday',
-                    'Saturday', 'Sunday']
-        return '%s %s to %s %s' % (weekdays[self.start_day],
-                                   self.start_time.strftime("%H:%M:%S"),
-                                   # to
-                                   weekdays[self.end_day],
-                                   self.end_time.strftime("%H:%M:%S"))
+        weekdays = [
+            "Monday",
+            "Tuesday",
+            "Wednesday",
+            "Thursday",
+            "Friday",
+            "Saturday",
+            "Sunday",
+        ]
+        return "%s %s to %s %s" % (
+            weekdays[self.start_day],
+            self.start_time.strftime("%H:%M:%S"),
+            # to
+            weekdays[self.end_day],
+            self.end_time.strftime("%H:%M:%S"),
+        )
+
 
 class Alert(TimeStampedModel):
     """
@@ -398,26 +400,30 @@ class Alert(TimeStampedModel):
 
     Alerts last for a period of time until the information is no longer dank.
     """
+
     # Define string constants to represent urgency tag levels
-    INFO = 'info'  # SRCT announcements
-    MINOR = 'minor'  # Holiday hours are in effect
-    MAJOR = 'major'  # The hungry patriot is closed today
-    EMERGENCY = 'emergency'  # Extreme weather
+    INFO = "info"  # SRCT announcements
+    MINOR = "minor"  # Holiday hours are in effect
+    MAJOR = "major"  # The hungry patriot is closed today
+    EMERGENCY = "emergency"  # Extreme weather
 
     # Tuple that ties a urgency tag with a string representation
     URGENCY_CHOICES = (
-        (INFO, 'Info'),
-        (MINOR, 'Minor'),
-        (MAJOR, 'Major'),
-        (EMERGENCY, 'Emergency'),
+        (INFO, "Advertising / Announcement"),
+        (MINOR, "Expected Hours Change"),
+        (MAJOR, "(Small Scale) Unexpected Hours Change"),
+        (EMERGENCY, "(University Wide) Unexpected Hours Change"),
     )
 
     # The urgency tag for this Alert
-    urgency_tag = models.CharField(max_length=10, default='Info',
-                                   choices=URGENCY_CHOICES)
+    urgency_tag = models.CharField(
+        max_length=10, default="Info", choices=URGENCY_CHOICES
+    )
 
     # The text that is displayed that describes the Alert
-    message = models.CharField(max_length=140)
+    subject = models.CharField(max_length=130)
+    body = models.TextField()
+    url = models.URLField("Reference URL", max_length=200)
 
     # The date + time that the alert will be start being served
     start_datetime = models.DateTimeField()
@@ -437,4 +443,4 @@ class Alert(TimeStampedModel):
         """
         String representation of an Alert object.
         """
-        return "%s" % (self.message)
+        return "{0} \n {1} \n {3}".format(self.subject, self.body, self.url)